+++ src/main/java/tornadofx/Async.kt
@@ -257,7 +257,7 @@ fun <T : Any> Node.runAsyncWithProgress(progress: Node = ProgressIndicator(), op
         val paddingVertical = (this as? Region)?.paddingVertical?.toDouble() ?: 0.0
         (progress as? Region)?.setPrefSize(boundsInParent.width - paddingHorizontal, boundsInParent.height - paddingVertical)
         // Unwrap ToolBar parent, it has an extra HBox or VBox inside it, we need to target the items list
-        val p = (parent?.parent as? ToolBar) ?: parent
+        val p = (parent?.parent as? ToolBar) ?: parent ?: throw IllegalArgumentException("runAsyncWithProgress cannot target an UI element with no parent!")
         val children = requireNotNull(p.getChildList()) { "This node has no child list, and cannot contain the progress node" }
         val index = children.indexOf(this)
         children.add(index, progress)