+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - The Workspace inside the `scope` of a UIComponents will assume the Workspace it is docked in (https://github.com/edvin/tornadofx/issues/806)
 - Kotlin 1.3.11
 - bindSelected for ViewModel gets `out` modifier (https://github.com/edvin/tornadofx/issues/823)
+- Spinner.required() validator (https://github.com/edvin/tornadofx/issues/871)
 
 ### Additions